refactor(ui): flatten `event-flow` directories
- `components/` had one file then a directory with other files
  - the `event-flow-details/` dir with several closely related files makes sense as an abstraction
  - but the `components/` dir itself with one file does not and is harder to read

- add new line between external and internal imports
