+++ ui/src/app/event-flow/event-flow-details/event-flow-page.scss
