+++ ui/src/app/event-flow/event-flow-details/genres.ts
