+++ ui/src/app/event-flow/event-flow-details/id.ts
