+++ ui/src/app/event-flow/index.ts
@@ -1,4 +1,4 @@
-import {EventFlowContainer} from './components/event-flow-container';
+import {EventFlowContainer} from './event-flow-container';
 
 export default {
     component: EventFlowContainer
